requirements.txt
Flask==2.0.2
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.4.1
requests==2.21.0
google-cloud-ndb==1.11.1